ぬるぽおじさんを作った

+++ Assets/kuroda/Scripts/context/context.cs
@@ -5,6 +5,5 @@ public class Context : MonoBehaviour
 {
     public Transform EnemyTransform { get; set; }
     public Transform PlayerTransform { get; set; }
-    public int speed { get; set; }
-
+    public float speed { get; set; }
 }
